Moved quick_resource and Fake helpers to test/helpers/resources.rb
